refactor: Remove uint usage (CLS) #8

+++ src/BloomFilter.CSRedis/FilterCSRedis.cs
@@ -1,7 +1,6 @@
 ﻿using CSRedis;
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -24,7 +23,7 @@ namespace BloomFilter.CSRedis
         /// <param name="expectedElements">The expected elements.</param>
         /// <param name="errorRate">The error rate.</param>
         /// <param name="hashFunction">The hash function.</param>
-        public FilterCSRedis(string name, CSRedisClient client, string redisKey, uint expectedElements, double errorRate, HashFunction hashFunction)
+        public FilterCSRedis(string name, CSRedisClient client, string redisKey, long expectedElements, double errorRate, HashFunction hashFunction)
             : base(name, expectedElements, errorRate, hashFunction)
         {
             if (string.IsNullOrWhiteSpace(redisKey)) throw new ArgumentException(nameof(redisKey));
@@ -41,7 +40,7 @@ namespace BloomFilter.CSRedis
         /// <param name="capacity">The capacity.</param>
         /// <param name="hashes">The hashes.</param>
         /// <param name="hashFunction">The hash function.</param>
-        public FilterCSRedis(string name, CSRedisClient client, string redisKey, uint capacity, uint hashes, HashFunction hashFunction)
+        public FilterCSRedis(string name, CSRedisClient client, string redisKey, long capacity, int hashes, HashFunction hashFunction)
             : base(name, capacity, hashes, hashFunction)
         {
             if (string.IsNullOrWhiteSpace(redisKey)) throw new ArgumentException(nameof(redisKey));
@@ -64,7 +63,7 @@ namespace BloomFilter.CSRedis
 
         public override IList<bool> Add(IEnumerable<byte[]> elements)
         {
-            var addHashs = new List<uint>();
+            var addHashs = new List<long>();
             foreach (var element in elements)
             {
                 addHashs.AddRange(ComputeHash(element));
@@ -90,7 +89,7 @@ namespace BloomFilter.CSRedis
 
         public override async ValueTask<IList<bool>> AddAsync(IEnumerable<byte[]> elements)
         {
-            var addHashs = new List<uint>();
+            var addHashs = new List<long>();
             foreach (var element in elements)
             {
                 addHashs.AddRange(ComputeHash(element));
@@ -129,7 +128,7 @@ namespace BloomFilter.CSRedis
 
         public override IList<bool> Contains(IEnumerable<byte[]> elements)
         {
-            var addHashs = new List<uint>();
+            var addHashs = new List<long>();
             foreach (var element in elements)
             {
                 addHashs.AddRange(ComputeHash(element));
@@ -156,7 +155,7 @@ namespace BloomFilter.CSRedis
 
         public override async ValueTask<IList<bool>> ContainsAsync(IEnumerable<byte[]> elements)
         {
-            var addHashs = new List<uint>();
+            var addHashs = new List<long>();
             foreach (var element in elements)
             {
                 addHashs.AddRange(ComputeHash(element));
@@ -206,49 +205,49 @@ namespace BloomFilter.CSRedis
             _client.Dispose();
         }
 
-        private IList<bool> SetBit(uint[] positions)
+        private IList<bool> SetBit(long[] positions)
         {
             var results = new bool[positions.Length];
 
             for (int i = 0; i < positions.Length; i++)
             {
-                results[i] = _client.SetBit(_redisKey, positions[i], true);
+                results[i] = _client.SetBit(_redisKey, (uint)positions[i], true);
             }
 
             return results;
         }
 
-        private async Task<IList<bool>> SetBitAsync(uint[] positions)
+        private async Task<IList<bool>> SetBitAsync(long[] positions)
         {
             var results = new bool[positions.Length];
 
             for (int i = 0; i < positions.Length; i++)
             {
-                results[i] = await _client.SetBitAsync(_redisKey, positions[i], true);
+                results[i] = await _client.SetBitAsync(_redisKey, (uint)positions[i], true);
             }
 
             return results;
         }
 
-        private IList<bool> GetBit(uint[] positions)
+        private IList<bool> GetBit(long[] positions)
         {
             var results = new bool[positions.Length];
 
             for (int i = 0; i < positions.Length; i++)
             {
-                results[i] = _client.GetBit(_redisKey, positions[i]);
+                results[i] = _client.GetBit(_redisKey, (uint)positions[i]);
             }
 
             return results;
         }
 
-        private async Task<IList<bool>> GetBitAsync(uint[] positions)
+        private async Task<IList<bool>> GetBitAsync(long[] positions)
         {
             var results = new bool[positions.Length];
 
             for (int i = 0; i < positions.Length; i++)
             {
-                results[i] = await _client.GetBitAsync(_redisKey, positions[i]);
+                results[i] = await _client.GetBitAsync(_redisKey, (uint)positions[i]);
             }
 
             return results;
