+++ src/BloomFilter.Redis/FilterRedisBuilder.cs
@@ -16,7 +16,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="expectedElements">The expected elements.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(string configuration, string redisKey, uint expectedElements, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(string configuration, string redisKey, long expectedElements, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(configuration, redisKey, expectedElements, 0.01, name);
     }
@@ -30,7 +30,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashMethod">The hash method.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(string configuration, string redisKey, uint expectedElements, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(string configuration, string redisKey, long expectedElements, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(configuration, redisKey, expectedElements, 0.01, hashMethod, name);
     }
@@ -44,7 +44,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashFunction">The hash function.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(string configuration, string redisKey, uint expectedElements, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(string configuration, string redisKey, long expectedElements, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(configuration, redisKey, expectedElements, 0.01, hashFunction, name);
     }
@@ -58,7 +58,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="errorRate">The error rate.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(string configuration, string redisKey, uint expectedElements, double errorRate, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(string configuration, string redisKey, long expectedElements, double errorRate, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(configuration, redisKey, expectedElements, errorRate, HashFunction.Functions[HashMethod.Murmur3], name);
     }
@@ -73,7 +73,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashMethod">The hash method.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(string configuration, string redisKey, uint expectedElements, double errorRate, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(string configuration, string redisKey, long expectedElements, double errorRate, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return new FilterRedis(name, new RedisBitOperate(configuration), redisKey, expectedElements, errorRate, HashFunction.Functions[hashMethod]);
     }
@@ -88,7 +88,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashFunction">The hash function.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(string configuration, string redisKey, uint expectedElements, double errorRate, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(string configuration, string redisKey, long expectedElements, double errorRate, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return new FilterRedis(name, new RedisBitOperate(configuration), redisKey, expectedElements, errorRate, hashFunction);
     }
@@ -101,7 +101,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="expectedElements">The expected elements.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(ConfigurationOptions configuration, string redisKey, uint expectedElements, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(ConfigurationOptions configuration, string redisKey, long expectedElements, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(configuration, redisKey, expectedElements, 0.01, name);
     }
@@ -115,7 +115,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashMethod">The hash method.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(ConfigurationOptions configuration, string redisKey, uint expectedElements, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(ConfigurationOptions configuration, string redisKey, long expectedElements, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(configuration, redisKey, expectedElements, 0.01, hashMethod, name);
     }
@@ -129,7 +129,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashFunction">The hash function.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(ConfigurationOptions configuration, string redisKey, uint expectedElements, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(ConfigurationOptions configuration, string redisKey, long expectedElements, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(configuration, redisKey, expectedElements, 0.01, hashFunction, name);
     }
@@ -143,7 +143,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="errorRate">The error rate.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(ConfigurationOptions configuration, string redisKey, uint expectedElements, double errorRate, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(ConfigurationOptions configuration, string redisKey, long expectedElements, double errorRate, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(configuration, redisKey, expectedElements, errorRate, HashFunction.Functions[HashMethod.Murmur3], name);
     }
@@ -158,7 +158,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashMethod">The hash method.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(ConfigurationOptions configuration, string redisKey, uint expectedElements, double errorRate, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(ConfigurationOptions configuration, string redisKey, long expectedElements, double errorRate, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return new FilterRedis(name, new RedisBitOperate(configuration), redisKey, expectedElements, errorRate, HashFunction.Functions[hashMethod]);
     }
@@ -173,7 +173,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashFunction">The hash function.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(ConfigurationOptions configuration, string redisKey, uint expectedElements, double errorRate, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(ConfigurationOptions configuration, string redisKey, long expectedElements, double errorRate, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return new FilterRedis(name, new RedisBitOperate(configuration), redisKey, expectedElements, errorRate, hashFunction);
     }
@@ -187,7 +187,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashMethod">The hash method.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(IConnectionMultiplexer connection, string redisKey, uint expectedElements, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(IConnectionMultiplexer connection, string redisKey, long expectedElements, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(connection, redisKey, expectedElements, 0.01, hashMethod, name);
     }
@@ -201,7 +201,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashFunction">The hash function.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(IConnectionMultiplexer connectionn, string redisKey, uint expectedElements, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(IConnectionMultiplexer connectionn, string redisKey, long expectedElements, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(connectionn, redisKey, expectedElements, 0.01, hashFunction, name);
     }
@@ -215,7 +215,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="errorRate">The error rate.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(IConnectionMultiplexer connection, string redisKey, uint expectedElements, double errorRate, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(IConnectionMultiplexer connection, string redisKey, long expectedElements, double errorRate, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(connection, redisKey, expectedElements, errorRate, HashFunction.Functions[HashMethod.Murmur3], name);
     }
@@ -228,7 +228,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="expectedElements">The expected elements.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(IConnectionMultiplexer connection, string redisKey, uint expectedElements, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(IConnectionMultiplexer connection, string redisKey, long expectedElements, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(connection, redisKey, expectedElements, 0.01, HashFunction.Functions[HashMethod.Murmur3], name);
     }
@@ -243,7 +243,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashMethod">The hash method.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(IConnectionMultiplexer connection, string redisKey, uint expectedElements, double errorRate, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(IConnectionMultiplexer connection, string redisKey, long expectedElements, double errorRate, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return new FilterRedis(name, new RedisBitOperate(connection), redisKey, expectedElements, errorRate, HashFunction.Functions[hashMethod]);
     }
@@ -258,7 +258,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashFunction">The hash function.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(IConnectionMultiplexer connection, string redisKey, uint expectedElements, double errorRate, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(IConnectionMultiplexer connection, string redisKey, long expectedElements, double errorRate, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return new FilterRedis(name, new RedisBitOperate(connection), redisKey, expectedElements, errorRate, hashFunction);
     }
@@ -272,7 +272,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashMethod">The hash method.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(IRedisBitOperate redisBitOperate, string redisKey, uint expectedElements, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(IRedisBitOperate redisBitOperate, string redisKey, long expectedElements, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(redisBitOperate, redisKey, expectedElements, 0.01, hashMethod, name);
     }
@@ -286,7 +286,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashFunction">The hash function.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(IRedisBitOperate redisBitOperate, string redisKey, uint expectedElements, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(IRedisBitOperate redisBitOperate, string redisKey, long expectedElements, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(redisBitOperate, redisKey, expectedElements, 0.01, hashFunction, name);
     }
@@ -300,7 +300,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="errorRate">The error rate.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(IRedisBitOperate redisBitOperate, string redisKey, uint expectedElements, double errorRate, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(IRedisBitOperate redisBitOperate, string redisKey, long expectedElements, double errorRate, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(redisBitOperate, redisKey, expectedElements, errorRate, HashFunction.Functions[HashMethod.Murmur3], name);
     }
@@ -313,7 +313,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="expectedElements">The expected elements.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(IRedisBitOperate redisBitOperate, string redisKey, uint expectedElements, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(IRedisBitOperate redisBitOperate, string redisKey, long expectedElements, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return Build(redisBitOperate, redisKey, expectedElements, 0.01, HashFunction.Functions[HashMethod.Murmur3], name);
     }
@@ -328,7 +328,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashMethod">The hash method.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(IRedisBitOperate redisBitOperate, string redisKey, uint expectedElements, double errorRate, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(IRedisBitOperate redisBitOperate, string redisKey, long expectedElements, double errorRate, HashMethod hashMethod, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return new FilterRedis(name, redisBitOperate, redisKey, expectedElements, errorRate, HashFunction.Functions[hashMethod]);
     }
@@ -343,7 +343,7 @@ public class FilterRedisBuilder : FilterBuilder
     /// <param name="hashFunction">The hash function.</param>
     /// <param name="name"></param>
     /// <returns></returns>
-    public static IBloomFilter Build(IRedisBitOperate redisBitOperate, string redisKey, uint expectedElements, double errorRate, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
+    public static IBloomFilter Build(IRedisBitOperate redisBitOperate, string redisKey, long expectedElements, double errorRate, HashFunction hashFunction, string name = BloomFilterConstValue.DefaultRedisName)
     {
         return new FilterRedis(name, redisBitOperate, redisKey, expectedElements, errorRate, hashFunction);
     }
